Crate cumulus_zombienet_sdk_helpers

Enums

BlockToCheck

Specifies which block should occupy a full core.

Functions

assert_blocks_are_being_finalized

Assert that finality has not stalled.

assert_finality_lag

assert_para_is_registered

Asserts that the given para_id is registered at the relay chain.

assert_para_throughput

assert_para_throughput_with

Like assert_para_throughput, but accepts a closure to validate each backed candidate receipt.

assert_relay_parent_offset

Asserts that parachain blocks have the correct relay parent offset. This also checks that the relay chain descendants do not contain any session changes.

assign_cores

Assigns the given cores to the given para_id.

create_runtime_upgrade_call

Creates a runtime upgrade call using Sudo::sudo(System::set_code_without_checks).

ensure_is_last_block_in_core

Checks if the specified block is the last block in a core.

ensure_is_only_block_in_core

Checks if the specified block occupies a full core.

find_core_info

Returns [CoreInfo] for the given parachain block.

submit_extrinsic_and_wait_for_finalization_success

Submits the given call as signed transaction and waits for its successful finalization.

submit_extrinsic_and_wait_for_finalization_success_with_timeout

Submits the given call as transaction and waits timeout_secs for it successful finalization.

submit_sudo_runtime_upgrade

Submit a runtime upgrade via sudo and verify it was scheduled.

submit_unsigned_extrinsic_and_wait_for_finalization_success

Submits the given call as unsigned transaction and waits for it successful finalization.

wait_for_first_session_change

Wait for the first block with a session change.

wait_for_nth_session_change

Wait for the first block with the Nth session change.

wait_for_pvf_prepare

Waits until every relaychain validator in network reports polkadot_pvf_prepare_concluded >= min_total_prepares.

wait_for_runtime_upgrade

Wait until a runtime upgrade has happened.